Refactor request listener to an exported reusable field
