+++ package.json
@@ -31,6 +31,7 @@
   },
   "homepage": "https://github.com/nilsroesel/skeidjs/wiki",
   "devDependencies": {
+    "@types/colors": "^1.2.1",
     "@types/node": "^10.1.2",
     "gulp": "^3.9.1",
     "gulp-typescript": "^4.0.2",
@@ -40,6 +41,7 @@
     "typescript": "^2.8.3"
   },
   "dependencies": {
+    "colors": "^1.3.0",
     "reflect-metadata": "^0.1.12"
   }
 }
