+++ src/interfaces/application.interface.ts
@@ -6,10 +6,10 @@ import { Middleware } from './index';
 export interface ApplicationConfig {
     contentType: string;
     server: {
-        port: number;
-        maxConnections: number;
-        timeout: number;
-        keepAliveTimeout: number;
+        port?: number;
+        maxConnections?: number;
+        timeout?: number;
+        keepAliveTimeout?: number;
         https?: HttpsOptions;
     }
     middleware?: Middleware;
