+++ src/util/internal.index.ts
@@ -5,3 +5,4 @@ export * from './route';
 export * from './router';
 export * from './injector';
 export * from './middleware.iterator';
+export * from './request-listener.factory';
